PowerShell project for creating a Desktop icon
This project was a work project. We use a series of network drives and my boss wanted a branch staff to easily run "gpupdate /force" when they needed. This project creates a Desktop icon that runs a batch file!

(Note: Changed the directory of the Storage for this file, typically copied from a Network Drive)